feat(ili9341): use lvgl pure Python driver
Prior to this commit, the `driver.init_ili9341()` function used the C extension
driver for the ILI9341 display, which required compiling a fork of the
[`lv_micropython`][lv-mpy] repo using custom compile flags to support the
M5Stack. Since this commit, several pull requests have been merged into
[`lv_micropython`][lv-mpy] and [`lv_micropython_bindings`][lv-mpy-bindings]
which make it possible to configure the pure-Python ILI9341 driver at run-time
to work with the M5Stack while using the stock [`lv_micropython`][lv-mpy] repo.

In this commit, deprecate the `driver.init_ili9341()` function and replace it
with the `driver.M5ili9341()` class, which sub-classes the pure-Python ILI9341
driver (from [`lv_micropython_bindings`][lv-mpy-bindings]) to set the default
parameters to work with the M5Stack ILI9341 display.

[lv-mpy]: https://github.com/littlevgl/lv_micropython
[lv-mpy-bindings]: https://github.com/littlevgl/lv_binding_micropython

BREAKING CHANGE: replace the `driver.init_ili9341()` function with the
    `driver.M5ili9341()` class.

    To migrate the code follow the example below:

    Before:

    ```python
    ili = init_ili9341()
    ```

    After:

    ```python
    disp = M5ili9341()
    ili = {'disp': disp, 'disp_drv': disp.disp_drv}
    ```