feat(examples): [WIP] adapt for m5stack
This commit modifies the `advanced_demo.py` to work with the m5stack display and button encoder.
Note that the user can cycle through the various tabs but there is currently no support for interacting
with the widgets on each page. Also, the animated chart seems to interfere with the button encoder
driver so the interface is not very responsive when the animated chart tab is selected.