feat(share-spi): support sharing the SPI bus
This commit introduces a factory function that chooses which display driver to
instantiate depending on the value of the `init_spi` flag. If "init_spi==True", it
instantiates the C module display driver (which has been compiled to add the
display to a pre-initialized SPI bus). If "init_spi==False", it instantiates the
pure/hybrid python display driver.